+++ src/react/atlascode/rovo-dev/rovoDevView.tsx
@@ -416,9 +416,9 @@ const RovoDevView: React.FC = () => {
                     });
                     break;
 
-                case RovoDevProviderMessageType.ReturnText:
                 case RovoDevProviderMessageType.CreatePRComplete:
                 case RovoDevProviderMessageType.GetCurrentBranchNameComplete:
+                case RovoDevProviderMessageType.CheckGitChangesComplete:
                     break; // This is handled elsewhere
 
                 default:
@@ -536,27 +536,6 @@ const RovoDevView: React.FC = () => {
         [postMessage],
     );
 
-    // Function to get the original text of a file for planning diff
-    const getOriginalText = useCallback(
-        async (filePath: string, range?: number[]) => {
-            const uniqueNonce = `${Math.random()}-${Date.now()}`; // Unique identifier for the request
-            const res = await postMessagePromise(
-                {
-                    type: RovoDevViewResponseType.GetOriginalText,
-                    filePath,
-                    range: range && range.length === 2 ? range : undefined,
-                    requestId: uniqueNonce, // Unique identifier for the request
-                },
-                RovoDevProviderMessageType.ReturnText,
-                1500,
-                uniqueNonce,
-            );
-
-            return res.text || '';
-        },
-        [postMessagePromise],
-    );
-
     const isRetryAfterErrorButtonEnabled = useCallback(
         (uid: string) => retryAfterErrorEnabled === uid,
         [retryAfterErrorEnabled],
@@ -594,7 +573,6 @@ const RovoDevView: React.FC = () => {
                     openFile,
                     isRetryAfterErrorButtonEnabled,
                     retryPromptAfterError,
-                    getOriginalText,
                 }}
                 messagingApi={{
                     postMessage,
@@ -617,16 +595,9 @@ const RovoDevView: React.FC = () => {
                     actionsEnabled={currentState === State.WaitingForPrompt}
                 />
                 <div className="prompt-container">
-                    {' '}
                     <PromptContextCollection
                         content={promptContextCollection}
                         readonly={false}
-                        onAddContext={async () => {
-                            postMessage({
-                                type: RovoDevViewResponseType.AddContext,
-                                currentContext: promptContextCollection,
-                            });
-                        }}
                         onRemoveContext={(item: RovoDevContextItem) => {
                             setPromptContextCollection((prev) => ({
                                 ...prev,