+++ src/react/atlascode/rovo-dev/rovoDevViewMessages.tsx
@@ -7,7 +7,6 @@ export const enum RovoDevViewResponseType {
     OpenFile = 'openFile',
     UndoFileChanges = 'undoFileChanges',
     KeepFileChanges = 'keepFileChanges',
-    GetOriginalText = 'getOriginalText',
     CreatePR = 'createPR',
     RetryPromptAfterError = 'retryPromptAfterError',
     GetCurrentBranchName = 'getCurrentBranchName',
@@ -16,6 +15,7 @@ export const enum RovoDevViewResponseType {
     ReportChangedFilesPanelShown = 'reportChangedFilesPanelShown',
     ReportChangesGitPushed = 'reportChangesGitPushed',
     ReportThinkingDrawerExpanded = 'reportThinkingDrawerExpanded',
+    CheckGitChanges = 'checkGitChanges',
 }
 
 export interface ModifiedFile {
@@ -29,7 +29,6 @@ export type RovoDevViewResponse =
     | ReducerAction<RovoDevViewResponseType.OpenFile, { filePath: string; tryShowDiff: boolean; range?: number[] }>
     | ReducerAction<RovoDevViewResponseType.UndoFileChanges, { files: ModifiedFile[] }>
     | ReducerAction<RovoDevViewResponseType.KeepFileChanges, { files: ModifiedFile[] }>
-    | ReducerAction<RovoDevViewResponseType.GetOriginalText, { filePath: string; range?: number[]; requestId: string }>
     | ReducerAction<RovoDevViewResponseType.CreatePR, { payload: { branchName: string; commitMessage: string } }>
     | ReducerAction<RovoDevViewResponseType.RetryPromptAfterError>
     | ReducerAction<RovoDevViewResponseType.GetCurrentBranchName>
@@ -37,4 +36,5 @@ export type RovoDevViewResponse =
     | ReducerAction<RovoDevViewResponseType.ForceUserFocusUpdate>
     | ReducerAction<RovoDevViewResponseType.ReportChangedFilesPanelShown, { filesCount: number }>
     | ReducerAction<RovoDevViewResponseType.ReportChangesGitPushed, { pullRequestCreated: boolean }>
-    | ReducerAction<RovoDevViewResponseType.ReportThinkingDrawerExpanded>;
+    | ReducerAction<RovoDevViewResponseType.ReportThinkingDrawerExpanded>
+    | ReducerAction<RovoDevViewResponseType.CheckGitChanges>;