+++ src/views/jira/treeViews/jiraAssignedWorkItemsViewProvider.ts
@@ -27,8 +27,8 @@ import { executeJqlQuery, JiraIssueNode, loginToJiraMessageNode, TreeViewIssue }
 
 const AssignedWorkItemsViewProviderId = AssignedJiraItemsViewId;
 
-const InitialCumulativeJqlFetchEventName = 'ui.cumulativeJqlFetch.render.lcp';
-const RefreshCumulativeJqlFetchEventName = 'ui.cumulativeJqlFetch.update.lcp';
+const InitialCumulativeJqlFetchEventName = 'ui.jira.jqlFetch.render.lcp';
+const RefreshCumulativeJqlFetchEventName = 'ui.jira.jqlFetch.update.lcp';
 
 export class AssignedWorkItemsViewProvider extends Disposable implements TreeDataProvider<TreeItem> {
     private static readonly _treeItemConfigureJiraMessage = loginToJiraMessageNode;
@@ -97,7 +97,7 @@ export class AssignedWorkItemsViewProvider extends Disposable implements TreeDat
         commands.executeCommand(`${AssignedWorkItemsViewProviderId}.focus`, {});
     }
 
-    public dispose(): void {
+    public override dispose(): void {
         this._disposable.dispose();
     }